chore: renamed components to follow camelcase convention
