+++ src/pages/ProjectDetailsPage.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useParams } from "react-router-dom";
 
 import ProjectTemplate from "../ProjectTemplate";
-import projectData from "../projectData";
+import projectData from "../components/projectData";
 import "./portfolio_page.css";
 
 const ProjectsDetailsPage = () => {
